some small fixes
